Created BaseOrder class. Added CancelledOrder and ReturnOrder. Refactored Order to extend BaseOrder.
