+++ lib/zinc.rb
@@ -2,8 +2,10 @@ require 'json'
 require 'rest_client'
 
 require 'zinc/order'
+require 'zinc/cancelled_order'
 require 'zinc/product_details'
 require 'zinc/product_prices'
+require 'zinc/returned_order'
 require 'zinc/errors/authentication_error'
 
 module Zinc
