+++ lib/zinc/returned_order.rb
@@ -1,0 +1,21 @@
+require 'zinc/base_order'
+
+module Zinc
+  class ReturnedOrder < BaseOrder
+
+    def self.get(id, params={})
+      o = ReturnedOrder.new
+      response = Zinc.request(:post, url+'/'+id+'/return', params)
+      if response
+        o.set_values(response)
+      end
+      o
+    end
+
+    def self.url
+      Zinc.url_base+'returns'
+    end
+
+    private
+  end
+end
